Moved all native lib tests to the main project MGroup.LinearAlgebra.Tests
