--- tests/MGroup.LinearAlgebra.Implementations.NativeWin64.Tests/MGroup.LinearAlgebra.Implementations.NativeWin64.Tests.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <StartupObject></StartupObject>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Compute.Bindings.IntelMKL" Version="0.1.5" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.11.1" />
-    <PackageReference Include="xunit" Version="2.9.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\MGroup.LinearAlgebra.Implementations.NativeWin64\MGroup.LinearAlgebra.Implementations.NativeWin64.csproj" />
-    <ProjectReference Include="..\MGroup.LinearAlgebra.Tests\MGroup.LinearAlgebra.Tests.csproj" />
-  </ItemGroup>
-
-</Project>
