--- tests/MGroup.LinearAlgebra.Implementations.NativeWin64.Tests/Program.cs
@@ -1,7 +1,0 @@
-namespace MGroup.LinearAlgebra.Implementations.NativeWin64.Tests
-{
-	internal class Program
-	{
-		
-	}
-}
